Doing SQL in Postgre
I am attaching several libraries to access postgreSQL. the Ruby and Python Versions are working fine, the DBLibrary has issues with the asynchronic nature of access. In theory, can be used in REST request, but I haven't tested it in such context. Feedback is welcome!, and if someone wants to beef them up, feel welcome to drop me a line